+++ packages/vendure-plugin-accept-blue/test/helpers/mocks.ts
@@ -1,6 +1,7 @@
 export const haydenSavedPaymentMethods = [
   {
     id: 14556,
+    customer_id: 181937,
     created_at: '2024-01-11T10:44:14.000Z',
     name: 'Hayden Zieme',
     payment_method_type: 'card',
@@ -13,6 +14,7 @@ export const haydenSavedPaymentMethods = [
   },
   {
     id: 14969,
+    customer_id: 181937,
     created_at: '2024-02-06T09:10:53.000Z',
     name: null,
     payment_method_type: 'card',
@@ -25,6 +27,7 @@ export const haydenSavedPaymentMethods = [
   },
   {
     id: 15012,
+    customer_id: 181937,
     created_at: '2024-02-06T22:14:07.000Z',
     name: 'Hayden Zieme',
     payment_method_type: 'check',
@@ -36,6 +39,7 @@ export const haydenSavedPaymentMethods = [
   },
   {
     id: 15019,
+    customer_id: 181937,
     created_at: '2024-02-07T07:48:20.000Z',
     name: 'Hayden Zieme',
     payment_method_type: 'check',
@@ -47,6 +51,7 @@ export const haydenSavedPaymentMethods = [
   },
   {
     id: 15020,
+    customer_id: 181937,
     created_at: '2024-02-07T08:00:23.000Z',
     name: 'Hayden Zieme',
     payment_method_type: 'check',
@@ -58,6 +63,7 @@ export const haydenSavedPaymentMethods = [
   },
   {
     id: 15021,
+    customer_id: 181937,
     created_at: '2024-02-07T08:01:32.000Z',
     name: 'Hiede',
     payment_method_type: 'check',
@@ -69,6 +75,7 @@ export const haydenSavedPaymentMethods = [
   },
   {
     id: 15022,
+    customer_id: 181937,
     created_at: '2024-02-07T08:03:25.000Z',
     name: 'Hayden',
     payment_method_type: 'check',
@@ -80,6 +87,7 @@ export const haydenSavedPaymentMethods = [
   },
   {
     id: 15023,
+    customer_id: 181937,
     created_at: '2024-02-07T08:04:22.000Z',
     name: 'Hayden',
     payment_method_type: 'check',
@@ -91,6 +99,7 @@ export const haydenSavedPaymentMethods = [
   },
   {
     id: 15024,
+    customer_id: 181937,
     created_at: '2024-02-07T08:05:06.000Z',
     name: 'Hayden',
     payment_method_type: 'check',
@@ -102,6 +111,7 @@ export const haydenSavedPaymentMethods = [
   },
   {
     id: 15035,
+    customer_id: 181937,
     created_at: '2024-02-07T10:50:37.000Z',
     name: 'Hayden Zieme',
     payment_method_type: 'check',
@@ -113,6 +123,7 @@ export const haydenSavedPaymentMethods = [
   },
   {
     id: 15036,
+    customer_id: 181937,
     created_at: '2024-02-07T22:14:07.000Z',
     name: 'Hayden Zieme',
     payment_method_type: 'check',